+++ packages/dubbo-node-test/src/badweather/unsupported-content-encoding.spec.ts
@@ -20,7 +20,7 @@ import {
   codeFromHttpStatus,
   endStreamFromJson,
   errorFromJsonBytes,
-} from "apache-dubbo/protocol-connect";
+} from "apache-dubbo/protocol-triple";
 import { http2Request } from "../helpers/http2-request.js";
 
 describe("unsupported content encoding", () => {
