+++ packages/dubbo/package.json
@@ -30,10 +30,10 @@
       "import": "./dist/esm/protocol/index.js",
       "require": "./dist/cjs/protocol/index.js"
     },
-    "./protocol-connect": {
-      "types": "./dist/types/protocol-connect/index.d.ts",
-      "import": "./dist/esm/protocol-connect/index.js",
-      "require": "./dist/cjs/protocol-connect/index.js"
+    "./protocol-triple": {
+      "types": "./dist/types/protocol-triple/index.d.ts",
+      "import": "./dist/esm/protocol-triple/index.js",
+      "require": "./dist/cjs/protocol-triple/index.js"
     },
     "./protocol-grpc": {
       "types": "./dist/types/protocol-grpc/index.d.ts",
@@ -51,8 +51,8 @@
       "protocol": [
         "./dist/types/protocol/index.d.ts"
       ],
-      "protocol-connect": [
-        "./dist/types/protocol-connect/index.d.ts"
+      "protocol-triple": [
+        "./dist/types/protocol-triple/index.d.ts"
       ],
       "protocol-grpc": [
         "./dist/types/protocol-grpc/index.d.ts"
