+++ packages/dubbo/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "files": [
     "src/index.ts",
-    "src/protocol-connect/index.ts",
+    "src/protocol-triple/index.ts",
     "src/protocol-grpc/index.ts",
     "src/protocol-grpc-web/index.ts"
   ],
